server/requirements.txt
openai-chatkit
fastapi
openai-agents
uvicorn